+++ src/main/java/webserver/controller/UserLoginController.java
@@ -3,8 +3,11 @@ package webserver.controller;
 import java.io.IOException;
 import java.net.URISyntaxException;
 
+import db.DataBase;
+import model.User;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
+import webserver.Parameters;
 
 public class UserLoginController extends AbstractController {
 
@@ -14,7 +17,23 @@ public class UserLoginController extends AbstractController {
     }
 
     @Override
-    protected void doPost(HttpRequest request, HttpResponse response) throws IOException, URISyntaxException {
+    protected void doPost(HttpRequest request, HttpResponse response) {
+        String requestBody = request.getRequestBody();
+        if (isLoggedIn(requestBody)) {
+            response.addHeader("Set-Cookie", "logined=true; Path=/");
+            response.sendRedirect("/index.html");
+            return;
+        }
+        response.addHeader("Set-Cookie", "logined=false");
+        response.sendRedirect("/user/login_failed.html");
+    }
 
+    private boolean isLoggedIn(String requestBody) {
+        Parameters parameters = new Parameters(requestBody);
+        String userId = parameters.getParameter("userId");
+        String password = parameters.getParameter("password");
+
+        User user = DataBase.findUserById(userId);
+        return user != null && password.equals(user.getPassword());
     }
 }
